+++ Modules/xNetworking/Examples/Resources/xNetAdapterAdvancedProperty/1-JumboPacket.ps1
@@ -1,0 +1,25 @@
+<#
+    .EXAMPLE
+    This configuration changes the JumboPacket Size.
+#>
+Configuration Example
+{
+    param
+    (
+        [Parameter()]
+        [System.String[]]
+        $NodeName = 'localhost'
+    )
+
+    Import-DSCResource -ModuleName xNetworking
+
+    Node $NodeName
+    {
+        xNetAdapterAdvancedProperty JumboPacket9014
+        {
+            NetworkAdapterName  = 'Ethernet'
+            RegistryKeyword     = "*JumboPacket"
+            RegistryValue       = 9014
+        }
+    }
+}